requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.13.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filer==2.1.1
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==5.2.0
django-polymorphic==3.0.0
django-prometheus==2.1.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
easy-thumbnails==2.8
et-xmlfile==1.1.0
flake8==3.9.2
kombu==5.1.0
lxml==4.6.4
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.31
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tablib==3.0.0
tinycss2==1.1.0
Unidecode==1.1.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
